删除操作时参数绑定

+++ src/main/java/com/wkclz/core/base/BaseService.java
@@ -1,6 +1,7 @@
 package com.wkclz.core.base;
 
 import com.wkclz.core.base.annotation.Desc;
+import com.wkclz.core.exception.BizException;
 import com.wkclz.core.helper.BaseHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -90,6 +91,9 @@ public class BaseService<Model extends BaseModel, Mapper extends BaseMapper<Mode
     public Integer delete(@NotNull Model model){
         BaseModel baseModel = new BaseModel();
         List<Long> ids = BaseHelper.getIdsFromBaseModel(model);
+        if (ids.isEmpty()) {
+            throw BizException.error("id or ids can not be null at the same time");
+        }
         baseModel.setIds(ids);
         return mapper.deleteBatch(model);
     }
